requirements.txt
flask
openai
python-dotenv
mysql-connector-python